requirements.txt
python-telegram-bot==22.5
python-dotenv==1.1.1
openai==2.5.0
chromadb==1.2.1
numpy==2.3.4